feat(core): extract shared runtime crate
