+++ bench_snapshots/rtx-5090/qwen3-4b.json
@@ -1,5 +1,5 @@
 {
-  "commit": "3448f87",
+  "commit": "59ca67c",
   "date": "2026-05-03",
   "model": "Qwen3-4B",
   "gpu": "NVIDIA GeForce RTX 5090",
@@ -8,21 +8,21 @@
     "output_len": 1,
     "metrics": {
       "ttft_ms": {
-        "avg_ms": 501.71845,
-        "p50_ms": 501.93214600000005,
-        "p95_ms": 503.511198,
-        "p99_ms": 503.749592,
-        "max_ms": 503.749592,
+        "avg_ms": 501.92715300000003,
+        "p50_ms": 502.59935,
+        "p95_ms": 503.73513199999996,
+        "p99_ms": 503.935338,
+        "max_ms": 503.935338,
         "samples": 20
       },
       "first_decode_step_ms": null,
       "steady_tpot_ms": null,
       "e2e_ms": {
-        "avg_ms": 501.718507,
-        "p50_ms": 501.932229,
-        "p95_ms": 503.511238,
-        "p99_ms": 503.74962700000003,
-        "max_ms": 503.74962700000003,
+        "avg_ms": 501.927236,
+        "p50_ms": 502.59940800000004,
+        "p95_ms": 503.735252,
+        "p99_ms": 503.93539899999996,
+        "max_ms": 503.93539899999996,
         "samples": 20
       },
       "generated_tokens": {
@@ -31,7 +31,7 @@
         "avg": 1.0,
         "samples": 20
       },
-      "request_tok_s": 1.9931495164107431,
+      "request_tok_s": 1.99232065522173,
       "decode_tok_s": null
     }
   },
@@ -40,35 +40,35 @@
     "output_len": 256,
     "metrics": {
       "ttft_ms": {
-        "avg_ms": 43.963086,
-        "p50_ms": 43.980075,
-        "p95_ms": 44.021534,
-        "p99_ms": 44.046838,
-        "max_ms": 44.046838,
+        "avg_ms": 43.924397,
+        "p50_ms": 43.927611,
+        "p95_ms": 43.993753,
+        "p99_ms": 44.05758,
+        "max_ms": 44.05758,
         "samples": 20
       },
       "first_decode_step_ms": {
-        "avg_ms": 7.373056999999999,
-        "p50_ms": 7.3730139999999995,
-        "p95_ms": 7.381652,
-        "p99_ms": 7.384898,
-        "max_ms": 7.384898,
+        "avg_ms": 7.365519,
+        "p50_ms": 7.365696,
+        "p95_ms": 7.384589,
+        "p99_ms": 7.388561,
+        "max_ms": 7.388561,
         "samples": 20
       },
       "steady_tpot_ms": {
-        "avg_ms": 7.398491,
-        "p50_ms": 7.395912,
-        "p95_ms": 7.456175,
-        "p99_ms": 7.462056,
-        "max_ms": 7.63168,
+        "avg_ms": 7.387237,
+        "p50_ms": 7.385105,
+        "p95_ms": 7.445011,
+        "p99_ms": 7.454451,
+        "max_ms": 8.212061,
         "samples": 5080
       },
       "e2e_ms": {
-        "avg_ms": 1930.5532210000001,
-        "p50_ms": 1930.6517190000002,
-        "p95_ms": 1930.9557260000001,
-        "p99_ms": 1931.115229,
-        "max_ms": 1931.115229,
+        "avg_ms": 1927.6482879999999,
+        "p50_ms": 1927.0942639999998,
+        "p95_ms": 1930.6249659999999,
+        "p99_ms": 1930.6948020000002,
+        "max_ms": 1930.6948020000002,
         "samples": 20
       },
       "generated_tokens": {
@@ -77,8 +77,8 @@
         "avg": 256.0,
         "samples": 20
       },
-      "request_tok_s": 132.60447688026656,
-      "decode_tok_s": 135.16450174336637
+      "request_tok_s": 132.80430956919426,
+      "decode_tok_s": 135.3701674957798
     }
   }
 }
