+++ crates/pegainfer-core/src/ffi.rs
@@ -1,0 +1,1 @@
+pub use pegainfer_kernels::ffi::*;
