+++ crates/pegainfer-core/src/lib.rs
@@ -1,0 +1,12 @@
+//! Shared runtime API used by pegainfer model crates.
+
+pub mod cuda_graph;
+pub mod ffi;
+pub mod kv_cache;
+pub mod kv_pool;
+pub mod model;
+pub mod ops;
+pub mod page_pool;
+pub mod sampler;
+pub mod tensor;
+pub mod weight_loader;
