+++ crates/pegainfer-core/src/ops/sampling.rs
