+++ crates/pegainfer-core/src/tensor.rs
@@ -1,0 +1,1 @@
+pub use pegainfer_kernels::tensor::*;
