+++ src/lib.rs
@@ -5,13 +5,10 @@ pub mod logging;
 pub mod model;
 pub(crate) mod model_executor;
 pub mod ops;
-#[allow(dead_code)]
-pub(crate) mod page_pool;
 pub mod sampler;
 pub mod scheduler;
 pub mod scheduler_qwen35;
 pub mod server_engine;
 pub mod tensor;
-pub mod trace_reporter;
 pub mod vllm_frontend;
 pub mod weight_loader;
